+++ .gitignore
@@ -6,4 +6,5 @@ build
 dist
 *.spec
 *.xlsx
+*.csv
 .DS_Store
