--- src/abrigados-entrada.py
@@ -1,6 +1,0 @@
-from deps import cadastroTerminal, PATH_ABR
-
-if __name__ == '__main__':
-    print('CADASTRO DE ABRIGADO')
-    while True:
-        cadastroTerminal(PATH_ABR, "abrigado")
