--- src/abrigados-saida.py
@@ -1,6 +1,0 @@
-from deps import liberaTerminal, PATH_ABR
-
-if __name__ == '__main__':
-    print('SAÍDA DE ABRIGADO')
-    while True:
-        liberaTerminal(PATH_ABR, 'abrigado')
