+++ src/gui-abrigados.py
@@ -1,4 +1,4 @@
-from gui import janelaDeControle
+from deps.gui import janelaDeControle
 
 if __name__ == "__main__":
     janelaDeControle("abrigado")
