+++ src/gui-saude.py
@@ -1,0 +1,4 @@
+from deps.gui import janelaDeControle
+
+if __name__ == "__main__":
+    janelaDeControle("voluntário da saude")
