--- src/voluntarios-entrada.py
@@ -1,6 +1,0 @@
-from deps import cadastroTerminal, PATH_VOL
-
-if __name__ == '__main__':
-    print('CADASTRO DE VOLUNTÁRIO')
-    while True:
-        cadastroTerminal(PATH_VOL, "voluntário")
