--- src/voluntarios-saida.py
@@ -1,6 +1,0 @@
-from deps import liberaTerminal, PATH_VOL
-
-if __name__ == '__main__':
-    print('SAÍDA DE VOLUNTÁRIO')
-    while True:
-        liberaTerminal(PATH_VOL, 'voluntário')
